moved tinymce routes to admin route
