+++ app/routes.py
@@ -167,69 +167,3 @@ def preview_email(user_id):
     messages, assessments = get_student_updates(user_id)
     return render_template("weekly_email.html", messages=messages, assessments=assessments)
 
-@main_bp.get("/tinymce-editor")
-@login_required
-def tinymce_editor():
-    """TinyMCE rich text editor page"""
-    return render_template("tinyMCE.html")
-
-@main_bp.post("/tinymce-editor")
-@login_required  
-def save_tinymce_content():
-    """Handle TinyMCE form submission"""
-    # Validate CSRF token
-    from flask_wtf.csrf import validate_csrf
-    try:
-        validate_csrf(request.form.get('csrf_token'))
-    except Exception:
-        flash("CSRF token validation failed. Please try again.", "error")
-        return redirect(url_for('main.tinymce_editor'))
-    
-    content = request.form.get('content', '')
-    title = request.form.get('title', 'Untitled')
-    degree_code = request.form.get('degree_code', 'GENERAL')
-    week_released = request.form.get('week_released', 1)
-    
-    # Handle new targeting fields
-    scheduled_at = request.form.get('scheduled_at', None)
-    degree_type_target = request.form.get('degree_type_target', None)
-    location_target = request.form.get('location_target', None)
-    stage_target = request.form.get('stage_target', None)
-    
-    # Convert scheduled_at to datetime if provided
-    scheduled_datetime = None
-    if scheduled_at:
-        try:
-            from datetime import datetime
-            scheduled_datetime = datetime.fromisoformat(scheduled_at)
-        except Exception:
-            flash("Invalid datetime format for scheduling.", "warning")
-    
-    # Create message object (you may want to save to database here)
-    message_data = {
-        'title': title,
-        'content': content,
-        'degree_code': degree_code,
-        'week_released': int(week_released) if week_released else 1,
-        'scheduled_at': scheduled_datetime,
-        'degree_type_target': degree_type_target if degree_type_target else None,
-        'location_target': location_target if location_target else None,
-        'stage_target': stage_target if stage_target else None
-    }
-    
-    # Process the content and redirect back
-    targeting_info = []
-    if degree_type_target:
-        targeting_info.append(f"Degree: {degree_type_target}")
-    if location_target:
-        targeting_info.append(f"Location: {location_target}")
-    if stage_target:
-        targeting_info.append(f"Stage: {stage_target}")
-    
-    targeting_str = " | ".join(targeting_info) if targeting_info else "All students"
-    schedule_str = f" | Scheduled: {scheduled_datetime}" if scheduled_datetime else ""
-    
-    flash(f"Content saved! Title: {title} | Targeting: {targeting_str}{schedule_str} | Content length: {len(content)} characters", "success")
-    current_app.logger.info(f"TinyMCE Message Data: {message_data}")
-    
-    return redirect(url_for('admin.admin_dashboard'))
