+++ app/routes_admin.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, redirect, url_for, flash, render_template, current_app, send_from_directory, session
 from .forms import ChangeRightForm, EmailEditor, DeleteAccountForm, AdminMessageForm, AdminReminderForm, SupportPostForm, SupportContactForm
 from .models import Right, Message, User, Reminder, SupportPost, SupportContact
-from .check import login_and_rights_required
+from .check import login_and_rights_required, login_required
 from . import db
 from werkzeug.utils import secure_filename
 from datetime import datetime
@@ -193,3 +193,70 @@ def admin_create_contact():
 @login_and_rights_required(1)
 def uploaded_file(filename):
     return send_from_directory(current_app.config.get('UPLOAD_FOLDER'), filename)
+
+@admin_bp.get("/tinymce-editor")
+@login_and_rights_required(1)
+def tinymce_editor():
+    """TinyMCE rich text editor page"""
+    return render_template("tinyMCE.html")
+
+@admin_bp.post("/tinymce-editor")
+@login_and_rights_required(1)
+def save_tinymce_content():
+    """Handle TinyMCE form submission"""
+    # Validate CSRF token
+    from flask_wtf.csrf import validate_csrf
+    try:
+        validate_csrf(request.form.get('csrf_token'))
+    except Exception:
+        flash("CSRF token validation failed. Please try again.", "error")
+        return redirect(url_for('main.admin_dashboard'))
+    
+    content = request.form.get('content', '')
+    title = request.form.get('title', 'Untitled')
+    degree_code = request.form.get('degree_code', 'GENERAL')
+    week_released = request.form.get('week_released', 1)
+    
+    # Handle new targeting fields
+    scheduled_at = request.form.get('scheduled_at', None)
+    degree_type_target = request.form.get('degree_type_target', None)
+    location_target = request.form.get('location_target', None)
+    stage_target = request.form.get('stage_target', None)
+    
+    # Convert scheduled_at to datetime if provided
+    scheduled_datetime = None
+    if scheduled_at:
+        try:
+            from datetime import datetime
+            scheduled_datetime = datetime.fromisoformat(scheduled_at)
+        except Exception:
+            flash("Invalid datetime format for scheduling.", "warning")
+    
+    # Create message object (you may want to save to database here)
+    message_data = {
+        'title': title,
+        'content': content,
+        'degree_code': degree_code,
+        'week_released': int(week_released) if week_released else 1,
+        'scheduled_at': scheduled_datetime,
+        'degree_type_target': degree_type_target if degree_type_target else None,
+        'location_target': location_target if location_target else None,
+        'stage_target': stage_target if stage_target else None
+    }
+    
+    # Process the content and redirect back
+    targeting_info = []
+    if degree_type_target:
+        targeting_info.append(f"Degree: {degree_type_target}")
+    if location_target:
+        targeting_info.append(f"Location: {location_target}")
+    if stage_target:
+        targeting_info.append(f"Stage: {stage_target}")
+    
+    targeting_str = " | ".join(targeting_info) if targeting_info else "All students"
+    schedule_str = f" | Scheduled: {scheduled_datetime}" if scheduled_datetime else ""
+    
+    flash(f"Content saved! Title: {title} | Targeting: {targeting_str}{schedule_str} | Content length: {len(content)} characters", "success")
+    current_app.logger.info(f"TinyMCE Message Data: {message_data}")
+    
+    return redirect(url_for('admin.admin_dashboard'))
